refactor(server): #125 — extract network/diagnostics handlers into routes/network.rs
